Refactored the code to join all the similar components to a single generic file, added the creators comment on every function, deleted the old files, and removed the @Preview functions.
