+++ core/designsystem/src/main/java/com/iteneum/designsystem/components/Card.kt
@@ -1,39 +1,48 @@
 package com.iteneum.designsystem.components
 
-import android.util.Log
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.CornerSize
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Add
 import androidx.compose.material.icons.outlined.Comment
 import androidx.compose.material.icons.outlined.Favorite
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateListOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.constraintlayout.compose.ConstraintLayout
+import coil.compose.rememberAsyncImagePainter
+import com.iteneum.designsystem.theme.Bianca
+import com.iteneum.designsystem.theme.Drab
 import com.iteneum.designsystem.theme.LPTypography
 
+/**
+ * Created [LpPostCard]
+ *
+ * @param modifier to modify an specific property of the card
+ * @param content as a high order function
+ *
+ * @author Daniel Roldan
+ */
 @Composable
 fun LpBasicCard(
     modifier: Modifier = Modifier,
-    content: @Composable() (ColumnScope.() -> Unit)
+    content: @Composable (ColumnScope.() -> Unit)
 ) {
     Card(
         modifier = modifier
@@ -161,7 +170,6 @@ fun LpPostCard(
  * @param icon The icon to be displayed
  * @param description Description to be displayed below the icon
  * @param onCardClick onClick event when card is clicked
- * @sample IconTextCardExample This is a usage example
  *
  * @author Jose Guadalupe Rivera
  */
@@ -207,28 +215,143 @@ fun LpIconTextCard(
     }
 }
 
-
-// this is an example of LpIconTextCard implementation
-@Preview
+/**
+ * [LPGenericElevatedCard] it's a card to show in the Dashboard screen
+ *
+ * @param title String to modify the name of the card
+ * @param description String to modify the description within the card
+ * @param onButtonClick High order function to assign the action that this card will have
+ *
+ * @author Jesus Lopez Gonzalez
+ */
 @Composable
-fun IconTextCardExample() {
-    val list = remember {
-        mutableStateListOf("Amenities", "x", "y", "gf")
-    }
-    LazyVerticalGrid(
-        modifier = Modifier.fillMaxSize(),
-        columns = GridCells.Fixed(2),
-        contentPadding = PaddingValues(0.dp)
+fun LPGenericElevatedCard(
+    title: String,
+    description: String,
+    buttonText: String,
+    onButtonClick: () -> Unit
+) {
+    val colors = MaterialTheme.colorScheme
+    Card(
+        colors = CardDefaults.cardColors(
+            containerColor =  Drab,
+        ),
+        shape = MaterialTheme.shapes.medium.copy(all = CornerSize(12.dp)),
+        modifier = Modifier
+            .width(312.dp)
+            .height(196.dp)
+            .padding(16.dp)
     ) {
-        items(list) { description ->
-            LpIconTextCard(
-                icon = Icons.Outlined.Add,
-                description = description
-            ) { card ->
-                // onClick: your functionality here
-                Log.e("tag", card) // show selected card message
+        Column(
+            modifier = Modifier.padding(start = 20.dp, top = 20.dp, end = 20.dp, bottom = 10.dp)
+        ) {
+            Text(
+                text = title,
+                style = LPTypography.titleMedium,
+                color = colors.onPrimary,
+            )
+            Spacer(modifier = Modifier.height(6.dp))
+            Text(
+                text = description,
+                style = LPTypography.bodyMedium,
+                color = colors.onPrimary,
+                maxLines = 3,
+                overflow = TextOverflow.Ellipsis,
+                textAlign = TextAlign.Justify,
+                fontSize = 14.sp
+            )
+            Spacer(modifier = Modifier.height(10.dp))
+            Spacer(modifier = Modifier.weight(1f))
+            Box(
+                modifier = Modifier.background(
+                    Color.White.copy(alpha = 0.1f),
+                    shape = MaterialTheme.shapes.medium.copy(all = CornerSize(12.dp))
+                )
+            ) {
+                TextButton(
+                    onClick = onButtonClick,
+                    colors = ButtonDefaults.textButtonColors(
+                        contentColor = colors.onPrimary
+                    )
+                ) {
+                    Text(text = buttonText)
+                }
             }
         }
     }
 }
 
+/**
+ * [LPGenericElevatedCardImage] it's a card to show in the Community List screen
+ *
+ * @param imageUrl String that will be a url to get the image from the internet
+ * @param title String to modify the name of the card
+ * @param description String to modify the description within the card
+ * @param onClick High order function to assign the action that this card will have
+ *
+ * @author Jesus Lopez Gonzalez
+ */
+@Composable
+fun LPGenericElevatedCardImage(
+    imageUrl: String,
+    title: String,
+    description: String,
+    onClick: () -> Unit
+) {
+    Card(
+        shape = MaterialTheme.shapes.medium.copy(all = CornerSize(12.dp)),
+        colors = CardDefaults.cardColors(
+            containerColor = Drab,
+        ),
+        modifier = Modifier.padding(10.dp)
+    ) {
+        Card(
+            colors = CardDefaults.cardColors(
+                containerColor = Bianca,
+            ),
+            shape = MaterialTheme.shapes.medium.copy(all = CornerSize(11.dp)),
+            modifier = Modifier
+                .width(177.dp)
+                .height(245.dp)
+                .clickable(onClick = onClick)
+                .padding(1.2.dp)
+        ) {
+            Column {
+                Image(
+                    painter = rememberAsyncImagePainter(imageUrl),
+                    contentDescription = null,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(169.dp),
+                    contentScale = ContentScale.Crop
+                )
+                Box(modifier = Modifier.padding(12.dp)) {
+                    ConstraintLayout {
+                        val (titleText, descriptionText) = createRefs()
+
+                        Text(
+                            text = title,
+                            style = MaterialTheme.typography.titleLarge,
+                            color = MaterialTheme.colorScheme.onSurface,
+                            maxLines = 1,
+                            overflow = TextOverflow.Ellipsis,
+                            modifier = Modifier.constrainAs(titleText) {
+                                top.linkTo(parent.top)
+                            }
+                        )
+                        Text(
+                            text = description,
+                            style = MaterialTheme.typography.bodyMedium,
+                            color = MaterialTheme.colorScheme.onSurface,
+                            maxLines = 1,
+                            overflow = TextOverflow.Ellipsis,
+                            modifier = Modifier.constrainAs(descriptionText) {
+                                top.linkTo(titleText.bottom)
+                            }
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
